changed of variable names
